pyproject.toml
[project]
name = "sweetspot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=1.10.0",
    "fastmcp>=2.2.6",
    "openai>=1.77.0",
    "flask>=3.1.0",
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "pymongo>=4.12.1",
]
